Added the 'ActionSynchronizer' class
The 'ActionSynchronizer' is a task that is run periodically to
synchronize the list of actions inside the two databases:
 - the server database
 - and the restful service database.
Each missing action in one database is added to it, so at the end, both
have the same list of actions.

IMPORTANT: Currently, this implementation doesn't handle the difference
of an action's description and value. So, if two actions have the same
id, they are considered to be equals, and the potentially differences of
description and value are left untouched in each database.

Added some thread-safe database methods to manipulate the actions.
Renamed the list of actions variable inside the database (actions ->
actionList).